Frontend

+++ frontend/src/components/MessageDetails.js
@@ -1,4 +1,5 @@
 import {useMessageContext} from '../hooks/useMessageContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
@@ -6,10 +7,18 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 const MessageDetails = ({message}) => {
 
     const {dispatch} = useMessageContext()
+    const {user} = useAuthContext()
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
+        
         const response = await fetch('/api/message/'+message._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
 
         const json = await response.json()
